requirements.txt
pycryptodome
websockets